requirements.txt
appdirs==1.4.4
asgiref==3.4.1
awsebcli==3.20.2
backports.entry-points-selectable==1.1.1
botocore==1.21.65
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
distlib==0.3.2
Django==3.2.9
django-bootstrap-modal-forms==2.2.0
django-environ==0.8.1
filelock==3.4.0
Flask==2.0.2
future==0.16.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
mongoengine==0.23.1
pathspec==0.5.9
platformdirs==2.4.0
pygame==2.0.1
pymongo==3.12.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.8
virtualenv==20.10.0
wcwidth==0.1.9
Werkzeug==2.0.2
whitenoise==5.3.0
